service-fichiers/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
minio==7.2.0
python-multipart==0.0.6
python-jose
httpx
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'
requests==2.31.0
